requirements.txt
pyyaml==6.0.2
numpy
torch
torchaudio
packaging
transformers
sentencepiece
sacremoses
